frontend/requirements.txt
whitenoise
dnspython
uwsgi
appnope==0.1.2
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
black==21.9b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
ConfigArgParse==1.5.2
decorator==5.1.0
Django==3.2.8
django-hosts==4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
filelock==3.3.0
idna==3.2
iniconfig==1.1.1
ipython==7.29.0
jedi==0.18.0
jsonschema==4.1.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
packaging==21.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
py==1.10.0
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-forge==18.6.0
pytojsonschema @ git+https://github.com/readevalprint/pytojsonschema.git@9ec4cdf0fbf6f0a820b020a99c832f0acfa15de9
pytz==2021.3
regex==2021.9.30
requests==2.26.0
requests-file==1.5.1
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
tldextract==3.1.2
toml==0.10.2
tomli==1.2.1
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.6
wcwidth==0.2.5
